+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 ## [v1.1.5.4]
 
 ### Added
-
+    - Seperated out process/system uptimes on server settings
 ### Changed
 
 ### Removed